module/requirements.txt
#
# Copyright 2020 IBM Corp.
# SPDX-License-Identifier: Apache-2.0
#
aiohttp==3.7.4.post0
aioitertools==0.8.0
async-timeout==3.0.1
attrs==21.2.0
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
curlify==2.2.1
Flask==1.1.2
Flask-RESTful==0.3.8
idna==3.2
jmespath==0.10.0
kafka-python==2.0.2
kubernetes==19.15.0
multidict==5.1.0
pandas==1.3.4
pycparser==2.20
PyJWT==2.0.1
PyYAML==5.4.1
requests
ruamel.yaml==0.16.12
s3fs==2021.8.0
s3transfer
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.12.1
yarl==1.6.3
